--no-index-fks option
By default, Datasette creates an index for every foreign key column that is
added using the --extract-column option.

For large tables, this can dramatically increase the size of the resulting
database file on disk. The new --no-index-fks option allows you to disable
this feature to save on file size.

Refs #24 which will allow you to explicitly list which columns SHOULD have
an index created.